refactored: shared modules and the way they get imported into docker
