+++ .gitignore
@@ -8,3 +8,5 @@ node_modules
 npm-debug.log
 
 tmp
+env
+__pycache__
