+++ kuwala-pipelines/common/js_utils/prettier.config.js
