+++ kuwala-pipelines/common/js_utils/src/errors/generalError/index.js
