+++ kuwala-pipelines/common/scripts/init.sh
