+++ kuwala-pipelines/google-poi/dockerfile
@@ -1,9 +1,9 @@
 FROM python:3.9.5
 
-COPY ./shared /opt/shared
+COPY ./common/python_utils /opt/common/python_utils
 COPY ./google-poi /opt/app
 
-ENV PYTHONPATH="/opt:/opt/app"
+ENV PYTHONPATH="/opt/common:/opt/app"
 
 WORKDIR /opt/app
 RUN pip install -r requirements.txt
