+++ kuwala-pipelines/osm-poi/src/app/app.js
@@ -6,7 +6,7 @@ const cors = require('cors');
 const mongoose = require('mongoose');
 const bodyParser = require('body-parser');
 const pino = require('express-pino-logger');
-const { Middlewares } = require('../../../shared/js');
+const { Middlewares } = require('../../../common/js_utils');
 const { geojson, radius } = require('./routes');
 
 const app = express();
