+++ kuwala-pipelines/population-density/src/app/app.js
@@ -4,7 +4,7 @@ const cors = require('cors');
 const mongoose = require('mongoose');
 const bodyParser = require('body-parser');
 const pino = require('express-pino-logger');
-const { Middlewares } = require('../../../shared/js');
+const { Middlewares } = require('../../../common/js_utils');
 const { cell, geojson, radius } = require('./routes');
 
 const app = express();
